feat: Implement comprehensive user, activity, category, and authentication management with API, infrastructure, and unit tests.

+++ THtracker.Application/UseCases/Users/DeleteUserUseCase.cs
@@ -13,6 +13,9 @@ public class DeleteUserUseCase
 
     public virtual async Task<bool> ExecuteAsync(Guid id)
     {
+        // Validación manual
+        if (id == Guid.Empty)
+            throw new Exception("El id de usuario es obligatorio.");
         return await _repository.DeleteAsync(id);
     }
 }
